backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pandas
google-adk
langfuse